+++ R/many_lm.R
@@ -61,8 +61,15 @@
 #' handled in each model separately by
 #' `lm()`.
 #'
-#' @param ... Additional arguments. To
+#' @param ... Additional arguments. For
+#' [many_lm()], these arguments will
 #' be passed to `lm()`.
+#' For the update method of the output
+#' of [many_lm()], these arguments will
+#' used to update the call to [many_lm()].
+#' For other methods, these
+#' arguments will be passed to methods
+#' if applicable.
 #'
 #' @author Shu Fai Cheung <https://orcid.org/0000-0002-9871-9448>
 #'
@@ -181,6 +188,358 @@ many_lm <- function(models,
     attr(out2, "call") <- match.call()
     out2
   }
+
+#' @importFrom stats getCall
+
+#' @details
+#' The `getCall` method of the output of
+#' [many_lm()] extracts the stored call.
+#'
+#' @return
+#' The `getCall` method returns a call
+#' stored in the output of [many_lm()].
+#'
+#' @param x For the `getCall` method,
+#' this is the output of [many_lm()].
+#'
+#' @rdname many_lm
+#' @export
+getCall <- function(
+  x,
+  ...
+) {
+  attr(x, "call")
+}
+
+#' @details
+#' The output of [many_lm()] has a update
+#' method.
+#'
+#' @return
+#' For the update method of the output
+#' of [many_lm()], if `evaluate` is `TRUE`,
+#' the updated output of [many_lm()]
+#' will be returned. If `evaluate` is
+#' `FALSE`, the updated call will be
+#' returned.
+#'
+#' @param object For methods of the output
+#' of [many_lm()],
+#' it is the output of [many_lm()].
+#'
+#' @param evaluate If `TRUE`, the
+#' updated call will be evaluated. If
+#' `FALSE`, the updated call will be
+#' returned.
+#'
+#' @rdname many_lm
+#' @export
+update.lm_list_lmhelprs <- function(
+  object,
+  ...,
+  evaluate = TRUE
+) {
+  # Adapted from update.default()
+  call <- attr(object, "call")
+  extras <- match.call(expand.dots = FALSE)$...
+  if (length(extras)) {
+    existing <- !is.na(match(names(extras), names(call)))
+    for (a in names(extras)[existing]) call[[a]] <- extras[[a]]
+    if (any(!existing)) {
+      call <- c(as.list(call), extras[!existing])
+      call <- as.call(call)
+    }
+  }
+  if (evaluate) {
+    return(eval(call, parent.frame()))
+  } else {
+    return(call)
+  }
+}
+
+#' @details
+#' The `coef` method extracts the
+#' regression coefficients in the
+#' `lavaan` style: `y ~ x`, `y`
+#' the response variable and `x` a term.
+#'
+#' @return
+#' The `coef` method returns a numeric
+#' vector of the coefficients.
+#'
+#' @param y For some methods,
+#' if set to a character vector,
+#' only the coefficients of the models
+#' of the response variables listed in
+#' `y` will be returned. If `NULL`, all
+#' coefficients will be returned.
+#'
+#' @rdname many_lm
+#' @export
+coef.lm_list_lmhelprs <- function(
+  object,
+  y = NULL,
+  ...
+) {
+  out0 <- lm_list_to_partable(
+    object,
+    keep_intercepts = TRUE
+  )
+  i <- out0$op %in% c("~", "~1")
+  out1 <- out0[i, , drop = FALSE]
+  out1$lavlabel <- paste0(
+    out1$lhs,
+    out1$op,
+    out1$rhs
+  )
+  vnames <- unique(out1$lhs)
+  j <- sapply(
+        vnames,
+        \(x) {
+          !any(out1[out1$lhs == x, "op", drop = TRUE] %in% "~")
+        })
+  ov_x <- vnames[j]
+  ov_y <- vnames[!j]
+  if (!is.null(y)) {
+    if (!is.null(y)) {
+      tmp <- y %in% ov_y
+      if (!all(tmp)) {
+        tmp2 <- y[!tmp]
+        stop("y not all in the model: ",
+            paste0(tmp2, collapse = ", "))
+      }
+    }
+    ov_y <- y
+  }
+  out1 <- out1[out1$lhs %in% ov_y, , drop = FALSE]
+  out2 <- out1$est
+  names(out2) <- out1$lavlabel
+  out2
+}
+
+#' @return
+#' The `vcov` method returns the
+#' variance-covariance matrix of the
+#' parameter estimates of the models.
+#'
+#' @details
+#' The `vcov` method is used to extract
+#' the variance-covariance matrices of
+#' the models.
+#'
+#' @param vcov_args A named list of
+#' arguments to be passed to [stats::vcov()]
+#' when computing the variance-covariance
+#' matrices
+#' of the regression coefficients.
+#' Default is `list()`, an empty list.
+#'
+#' @rdname many_lm
+#' @export
+vcov.lm_list_lmhelprs <- function(
+  object,
+  y = NULL,
+  vcov_args = list(),
+  ...
+) {
+  est <- stats::coef(object, y = y, ...)
+  pvalue_fun <- NULL
+
+  # VCOV
+  vcov_all <- lapply(
+    object,
+    \(x) {
+      do.call(
+        stats::vcov,
+        c(list(object = x),
+          vcov_args)
+      )
+    }
+  )
+  y_names <- sapply(
+    object,
+    get_response
+  )
+  for (i in seq_along(vcov_all)) {
+    y_name <-y_names[i]
+    tmp1 <- vcov_all[[i]]
+    tmp2 <- colnames(tmp1)
+    tmp2 <- gsub("(Intercept)",
+                 "1",
+                 tmp2,
+                 fixed = TRUE)
+    tmp2 <- paste0(y_name, "~", tmp2)
+    colnames(tmp1) <- rownames(tmp1) <- tmp2
+    vcov_all[[i]] <- tmp1
+  }
+  p <- length(est)
+  out <- matrix(0, p, p)
+  est_names <-names(est)
+  colnames(out) <- rownames(out) <- est_names
+  for (vcov_i in vcov_all) {
+    tmp <- intersect(est_names, colnames(vcov_i))
+    if (length(tmp) < 1) next
+    out[tmp, tmp] <- vcov_i[tmp, tmp]
+  }
+  out
+}
+
+#' @details
+#' The `confint` method computes the
+#' confidence intervals for the coefficients.
+#'
+#' @return
+#' The `confint` method returns a two-column
+#' matrix of the confidence intervals.
+#'
+#' @param ci_fun The function to be used
+#' to form the confidence intervals for
+#' regression coefficients. Default
+#' is `stats::confint`
+#'
+#' @param ci_args A named list of
+#' arguments to be passed to `ci_fun`.
+#' Default is `list(level = .95)`,
+#' requesting 95% confidence intervals.
+#'
+#'
+#' @param parm The parameters for which
+#' the confidence intervals will be returned.
+#' If `NULL`, the confidence intervals
+#' for all coefficients will be returned.
+#'
+#' @param level The level of confidence
+#' of the confidence levels.
+#'
+#' @rdname many_lm
+#' @export
+confint.lm_list_lmhelprs <- function(
+  object,
+  parm = NULL,
+  level = .95,
+  ci_fun = stats::confint,
+  ci_args = list(),
+  ...
+) {
+  ci_args <- utils::modifyList(
+    ci_args,
+    list(level = level)
+  )
+  out0 <- lm_list_to_partable(
+    object = object,
+    keep_intercepts = TRUE,
+    ci = TRUE,
+    ci_fun = ci_fun,
+    ci_args = ci_args
+  )
+  est <- stats::coef(object)
+  out0$lavlabel <- paste0(
+    out0$lhs,
+    out0$op,
+    out0$rhs
+  )
+  out0 <- out0[out0$lavlabel %in% names(est), ]
+  tmp <- do.call(
+    ci_fun,
+    c(list(object = object[[1]]),
+      ci_args)
+  )
+  out1 <- out0[, c("ci.lower", "ci.upper")]
+  out1 <- as.matrix(out1)
+  rownames(out1) <- out0$lavlabel
+  out1 <- out1[names(est), ]
+  if (!is.null(parm)) {
+    out1 <- out1[parm, ]
+  }
+  colnames(out1) <- colnames(tmp)
+  out1
+}
+
+#' @importFrom stats nobs
+
+#' @details
+#' The `nobs` method extracts the
+#' number of cases actually used in
+#' the analysis. Should be identical
+#' for all models.
+#'
+#' @return
+#' The `nobs` method returns the
+#' number of cases used in the models.
+#'
+#' @rdname many_lm
+#' @export
+nobs.lm_list_lmhelprs <- function(
+  object,
+  ...
+) {
+  ns <- sapply(
+    object,
+    stats::nobs,
+  )
+  if (all(ns != ns[1])) {
+    stop("Something's wrong. The models do not have the same number of cases.")
+  }
+  ns[1]
+}
+
+
+
+#' @details
+#' Unlike the corresponding method for
+#' the output of [stats::lm()],
+#' the `model.frame` method of the output
+#' of [many_lm()] does only one thing:
+#' extracts the stored model frame. It is
+#' because [many_lm()] is for fitting
+#' several models, with several formulas
+#' instead of only one.
+#'
+#' @return
+#' The `model.frame` method of the output
+#' of [many_lm()] returns a model frame,
+#' formed by merging the model frames of
+#' all models.
+#'
+#' @param formula For the `model.frame`
+#' method, this should be the output
+#' of [many_lm()].
+#'
+#' @rdname many_lm
+#' @export
+model.frame.lm_list_lmhelprs <- function(
+  formula,
+  ...
+) {
+  out <- merge_model_matrices(formula)
+  out
+}
+
+#' @importFrom stats variable.names
+
+#' @details
+#' The `variable.names` method of the
+#' output of [many_lm()] extract the
+#' variable names of all models.
+#'
+#' @return
+#' The `variable.names` method returns
+#' a character vector of unique variable
+#' names in all the stored models.
+#'
+#' @rdname many_lm
+#' @export
+variable.names.lm_list_lmhelprs <- function(
+  object,
+  ...
+) {
+  out0 <- lapply(
+    object,
+    stats::variable.names
+  )
+  out1 <- unique(unlist(out0))
+  out1
+}
 
 #' @noRd