+++ inst/tinytest/test_test_many_lm_model_frame.R
@@ -1,0 +1,26 @@
+library(lmhelprs)
+
+dat <- data_test1
+dat[1:5, "x2"] <- NA
+dat[3:6, "y"] <- NA
+
+mod1 <- "
+
+# Comments
+         y ~ x2 + x1 + x3 + x4
+x3 ~ cat1 + x1 + x2*x4
+  # The last line
+
+
+         "
+
+out <- many_lm(
+  mod1,
+  dat
+)
+
+out_mm <- model.frame(out)
+
+expect_true(
+  all(c("x2:x4", "cat1Beta") %in% colnames(out_mm))
+)